requirements.txt
dacite==1.6.0
PyYAML==6.0.1
SQLAlchemy-Utils==0.37.8
SQLAlchemy==1.4.20
psycopg2-binary<=3.0.0; sys_platform == 'darwin'
psycopg2<=3.0.0; sys_platform == 'linux'
alembic==1.6.5
pytz==2024.1
click==8.1.3
lxml<=6.0.0
pandas>=2.1.1
fastapi==0.115.0
uvicorn>=0.11.7
pydantic<3.0.0
python-multipart==0.0.5
gunicorn==20.1.0
beautifulsoup4==4.12.3
py-consul==1.5.3
openai==1.51.2
langchain_openai==0.2.2
langchain_core==0.3.10
langchain==0.3.3
